src/main/java/dev/andrylat/carsharing/controllers/BodyTypesController.java
package dev.andrylat.carsharing.controllers;

import dev.andrylat.carsharing.services.BodyTypeService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;

@Controller
@RequestMapping("/bodytypes")
public class BodyTypesController {

    private final BodyTypeService bodyTypeService;

    @Autowired
    public BodyTypesController(BodyTypeService bodyTypeService) {
        this.bodyTypeService = bodyTypeService;
    }

    @GetMapping()
    public String getAll(@RequestParam(name = "pageNumber", defaultValue = "0") int pageNumber,
                         @RequestParam(name = "pageSize", defaultValue = "10") int pageSize,
                         Model model) {
        model.addAttribute("bodyTypes", bodyTypeService.getAll(pageNumber, pageSize));
        model.addAttribute("recordsNumber", bodyTypeService.getRecordsNumber());
        model.addAttribute("pageNumber", pageNumber);
        model.addAttribute("pageSize", pageSize);
        return "bodytypes/getAll";
    }

    @GetMapping("/{id}")
    public String getById(@PathVariable("id") long id, Model model) {
        model.addAttribute("bodyType", bodyTypeService.getById(id));
        return "bodytypes/getById";
    }

}